requirements/base.txt
# django version 1.11.10
django==2.1

# django rest framework
djangorestframework==3.8.2

# rest jwt
djangorestframework-jwt==1.11.0

# drf docs
drfdocs

# pillow for images
Pillow==5.2.0

# django allauth for user authentication
django-allauth==0.36.0

# django rest authentications
django-rest-auth==0.9.3

# django crispy forms 
django-crispy-forms==1.7.2

# django environ
django-environ==0.4.4

# postgresql file
psycopg2-binary==2.7.5

# authentications mixins
django-braces==1.13.0

# django cors
django-cors-headers==2.2.0

# python decouple for environmental variables
python-decouple==3.1

# Ipython for the python shell
ipython

# testing packages
pytest

# twilio
twilio==6.19.2

#open py excel
openpyxl==2.6.1